refactor: move suggested filters component to app router
