+++ src/app/(core)/teachers/search/useSuggestedFilters.test.tsx
@@ -1,7 +1,8 @@
 import { renderHook } from "@testing-library/react";
 
+import { SuggestedFilters } from "../../../../context/Search/search.types";
+
 import { useSuggestedFilters } from "./useSuggestedFilters";
-import { SuggestedFilters } from "./search.types";
 
 const mockUseSWR = jest.fn<
   { data: unknown; error: unknown; isLoading: boolean },
